Switch from gert to git2r
PROBLEM:
gert is not state-preserving. It would likely work butit relies on
git_branch_checkout() and git will not allow deletion of the
branch checked out. If the initial HEAD is a branch, this does
not pose a problem, but if the initial HEAD is a tag or commit,
a new branch created by giterateR will be added to the repo.